+++ example_runner.py
@@ -20,6 +20,7 @@ import example.chiptower_readme_example
 print('Can Tower Examples')
 import example.can_platform
 import example.can_tower
+import example.can_tower_set
 
 print('Walkway examples')
 import example.walkway